Create cluster_nodes.c for node serialization and config persistence
Extract cluster node serialization and nodes.conf persistence into a new
file cluster_nodes.c with its header cluster_nodes.h.

Moved from cluster.c:
- Aux field handlers (types, table, getter/setter functions, now static)
- representClusterNodeFlags, representSlotInfo
- clusterGenNodeDescription, clusterGenNodesDescription
- clusterGenNodesSlotsInfo, clusterFreeNodesSlotsInfo

Moved from cluster_legacy.c:
- clusterLoadConfig, clusterSaveConfig, clusterSaveConfigOrDie,
  clusterSaveConfigOrLog, clusterLockConfig

The load/save code now uses vtable callbacks instead of LEGACY_DATA/
LEGACY_STATE macros:
- parseVarsLine/appendVarsLine for the nodes.conf vars line
- setNodePingPongEpoch for per-node timing and epoch
- setNodeFailed for marking nodes as failed during load
- postLoad for post-load fixups

Also add clusterAddNode and clusterGetNodesInMyShard declarations to
cluster_state.h.
