+++ src/Makefile
@@ -465,6 +465,7 @@ ENGINE_SERVER_OBJ = \
     cluster.o \
     cluster_legacy.o \
     cluster_link.o \
+    cluster_nodes.o \
     cluster_state.o \
     cluster_migrateslots.o \
     cluster_slot_stats.o \
