+++ src/cluster.h
@@ -80,33 +80,6 @@ void clusterBroadcastPong(int target);
 unsigned long getClusterConnectionsCount(void);
 int isClusterHealthy(void);
 
-sds clusterGenNodesDescription(client *c, int filter, int tls_primary);
-sds representClusterNodeFlags(sds ci, uint16_t flags);
-
-/* Aux field handler types and table for nodes.conf serialization. */
-typedef int(aux_value_setter)(clusterNode *n, void *value, size_t length);
-typedef sds(aux_value_getter)(clusterNode *n, sds s);
-typedef int(aux_value_present)(clusterNode *n);
-typedef struct {
-    char *field;
-    aux_value_setter *setter;
-    aux_value_getter *getter;
-    aux_value_present *isPresent;
-} auxFieldHandler;
-typedef enum {
-    af_shard_id,
-    af_human_nodename,
-    af_tcp_port,
-    af_tls_port,
-    af_announce_client_ipv4,
-    af_announce_client_ipv6,
-    af_announce_client_tcp_port,
-    af_announce_client_tls_port,
-    af_availability_zone,
-    af_count,
-} auxFieldIndex;
-extern auxFieldHandler auxFieldHandlers[];
-
 sds genClusterInfoString(sds info);
 /* handle implementation specific debug cluster commands. Return 1 if handled, 0 otherwise. */
 int handleDebugClusterCommand(client *c);
@@ -119,9 +92,6 @@ void clusterCommandSlots(client *c);
 void clusterCommandMyId(client *c);
 void clusterCommandMyShardId(client *c);
 void clusterCommandShards(client *c);
-void clusterGenNodesSlotsInfo(int filter);
-void clusterFreeNodesSlotsInfo(clusterNode *n);
-sds clusterGenNodeDescription(client *c, clusterNode *node, int tls_primary);
 
 int clusterNodeCoversSlot(clusterNode *n, int slot);
 int getNodeDefaultClientPort(clusterNode *n);
@@ -184,8 +154,6 @@ unsigned int delKeysInSlot(unsigned int hashslot, int lazy, bool propagate_del, 
 
 unsigned int propagateSlotDeletionByKeys(unsigned int hashslot);
 void clusterUpdateState(void);
-void clusterSaveConfigOrDie(int do_fsync);
-int clusterSaveConfig(int do_fsync);
 int clusterBumpConfigEpochWithoutConsensus(void);
 void clusterDoBeforeSleep(int flags);
