+++ src/cluster_legacy.c
@@ -38,6 +38,7 @@
 
 #include "server.h"
 #include "cluster.h"
+#include "cluster_nodes.h"
 #include "cluster_state.h"
 #include "cluster_link.h"
 
@@ -600,563 +601,6 @@ static clusterMsg *getMessageFromSendBlock(clusterMsgSendBlock *block) {
  * Initialization
  * -------------------------------------------------------------------------- */
 
-/* Load the cluster config from 'filename'.
- *
- * If the file does not exist or is zero-length (this may happen because
- * when we lock the nodes.conf file, we create a zero-length one for the
- * sake of locking if it does not already exist), C_ERR is returned.
- * If the configuration was loaded from the file, C_OK is returned. */
-int clusterLoadConfig(char *filename) {
-    FILE *fp = fopen(filename, "r");
-    struct stat sb;
-    char *line;
-    int maxline, j;
-    dict *tmp_cluster_nodes;
-
-    if (fp == NULL) {
-        if (errno == ENOENT) {
-            return C_ERR;
-        } else {
-            serverLog(LL_WARNING, "Loading the cluster node config from %s: %s", filename, strerror(errno));
-            exit(1);
-        }
-    }
-
-    if (valkey_fstat(fileno(fp), &sb) == -1) {
-        serverLog(LL_WARNING, "Unable to obtain the cluster node config file stat %s: %s", filename, strerror(errno));
-        exit(1);
-    }
-    /* Check if the file is zero-length: if so return C_ERR to signal
-     * we have to write the config. */
-    if (sb.st_size == 0) {
-        fclose(fp);
-        return C_ERR;
-    }
-
-    /* Parse the file. Note that single lines of the cluster config file can
-     * be really long as they include all the hash slots of the node.
-     * This means in the worst possible case, half of the slots will be
-     * present in a single line, possibly in importing or migrating state, so
-     * together with the node ID of the sender/receiver.
-     *
-     * To simplify we allocate 1024+CLUSTER_SLOTS*128 bytes per line. */
-    maxline = 1024 + CLUSTER_SLOTS * 128;
-    line = zmalloc(maxline);
-    tmp_cluster_nodes = dictCreate(&clusterNodesDictType);
-    while (fgets(line, maxline, fp) != NULL) {
-        int argc, aux_argc;
-        sds *argv, *aux_argv;
-        clusterNode *n, *primary;
-        char *p, *s;
-
-        /* Skip blank lines, they can be created either by users manually
-         * editing nodes.conf or by the config writing process if stopped
-         * before the truncate() call. */
-        if (line[0] == '\n' || line[0] == '\0') continue;
-
-        /* Split the line into arguments for processing. */
-        argv = sdssplitargs(line, &argc);
-        if (argv == NULL) goto fmterr;
-
-        /* Handle the special "vars" line. Don't pretend it is the last
-         * line even if it actually is when generated by the server. */
-        if (strcasecmp(argv[0], "vars") == 0) {
-            if (!(argc % 2)) goto fmterr;
-            for (j = 1; j < argc; j += 2) {
-                if (strcasecmp(argv[j], "currentEpoch") == 0) {
-                    LEGACY_STATE()->currentEpoch = strtoull(argv[j + 1], NULL, 10);
-                } else if (strcasecmp(argv[j], "lastVoteEpoch") == 0) {
-                    LEGACY_STATE()->lastVoteEpoch = strtoull(argv[j + 1], NULL, 10);
-                } else {
-                    serverLog(LL_NOTICE, "Skipping unknown cluster config variable '%s'", argv[j]);
-                }
-            }
-            sdsfreesplitres(argv, argc);
-            continue;
-        }
-
-        /* Regular config lines have at least eight fields */
-        if (argc < 8) {
-            sdsfreesplitres(argv, argc);
-            goto fmterr;
-        }
-
-        /* Create this node if it does not exist */
-        if (verifyClusterNodeId(argv[0], sdslen(argv[0])) == C_ERR) {
-            sdsfreesplitres(argv, argc);
-            goto fmterr;
-        }
-        n = clusterLookupNode(argv[0], sdslen(argv[0]));
-        if (!n) {
-            n = createClusterNode(argv[0], 0);
-            clusterAddNode(n);
-            dictAdd(tmp_cluster_nodes, sdsnewlen(argv[0], sdslen(argv[0])), NULL);
-        } else {
-            /* Check if the node (nodeid) has already been loaded. The nodeid is used to
-             * identify every node across the entire cluster, we do not expect to find
-             * duplicate nodeids in nodes.conf. */
-            dictEntry *de = dictFind(tmp_cluster_nodes, argv[0]);
-            if (de != NULL) {
-                serverLog(LL_WARNING, "Duplicate nodeid detected: %s", argv[0]);
-                sdsfreesplitres(argv, argc);
-                goto fmterr;
-            }
-        }
-        /* Format for the node address and auxiliary argument information:
-         * ip:port[@cport][,hostname][,aux=val]*] */
-
-        aux_argv = sdssplitlen(argv[1], sdslen(argv[1]), ",", 1, &aux_argc);
-        if (aux_argv == NULL) {
-            sdsfreesplitres(argv, argc);
-            goto fmterr;
-        }
-
-        /* Hostname is an optional argument that defines the endpoint
-         * that can be reported to clients instead of IP. */
-        if (aux_argc > 1 && sdslen(aux_argv[1]) > 0) {
-            n->hostname = sdscpy(n->hostname, aux_argv[1]);
-        } else if (sdslen(n->hostname) != 0) {
-            sdsclear(n->hostname);
-        }
-
-        /* All fields after hostname are auxiliary and they take on
-         * the format of "aux=val" where both aux and val can contain
-         * characters that pass the isValidAuxChar check only. The order
-         * of the aux fields is insignificant. */
-        int aux_tcp_port = 0;
-        int aux_tls_port = 0;
-        for (int i = 2; i < aux_argc; i++) {
-            int field_argc;
-            sds *field_argv;
-            field_argv = sdssplitlen(aux_argv[i], sdslen(aux_argv[i]), "=", 1, &field_argc);
-            if (field_argv == NULL || field_argc != 2) {
-                /* Invalid aux field format */
-                if (field_argv != NULL) sdsfreesplitres(field_argv, field_argc);
-                sdsfreesplitres(aux_argv, aux_argc);
-                sdsfreesplitres(argv, argc);
-                goto fmterr;
-            }
-
-            /* Validate that both aux and value contain valid characters only */
-            for (unsigned j = 0; j < 2; j++) {
-                if (!isValidAuxString(field_argv[j], sdslen(field_argv[j]))) {
-                    /* Invalid aux field format */
-                    sdsfreesplitres(field_argv, field_argc);
-                    sdsfreesplitres(aux_argv, aux_argc);
-                    sdsfreesplitres(argv, argc);
-                    goto fmterr;
-                }
-            }
-
-            /* Note that we don't expect lots of aux fields in the foreseeable
-             * future so a linear search is completely fine. */
-            int field_found = 0;
-            for (unsigned j = 0; j < af_count; j++) {
-                if (sdslen(field_argv[0]) != strlen(auxFieldHandlers[j].field) ||
-                    memcmp(field_argv[0], auxFieldHandlers[j].field, sdslen(field_argv[0])) != 0) {
-                    continue;
-                }
-                field_found = 1;
-                aux_tcp_port |= j == af_tcp_port;
-                aux_tls_port |= j == af_tls_port;
-                if (auxFieldHandlers[j].setter(n, field_argv[1], sdslen(field_argv[1])) != C_OK) {
-                    /* Invalid aux field format */
-                    sdsfreesplitres(field_argv, field_argc);
-                    sdsfreesplitres(aux_argv, aux_argc);
-                    sdsfreesplitres(argv, argc);
-                    goto fmterr;
-                }
-            }
-
-            if (field_found == 0) {
-                /* Invalid aux field format */
-                sdsfreesplitres(field_argv, field_argc);
-                sdsfreesplitres(aux_argv, aux_argc);
-                sdsfreesplitres(argv, argc);
-                goto fmterr;
-            }
-
-            sdsfreesplitres(field_argv, field_argc);
-        }
-        /* Address and port */
-        if ((p = strrchr(aux_argv[0], ':')) == NULL) {
-            sdsfreesplitres(aux_argv, aux_argc);
-            sdsfreesplitres(argv, argc);
-            goto fmterr;
-        }
-        *p = '\0';
-        memcpy(n->ip, aux_argv[0], strlen(aux_argv[0]) + 1);
-        char *port = p + 1;
-        char *busp = strchr(port, '@');
-        if (busp) {
-            *busp = '\0';
-            busp++;
-        }
-        /* If neither TCP or TLS port is found in aux field, it is considered
-         * an old version of nodes.conf file.*/
-        if (!aux_tcp_port && !aux_tls_port) {
-            if (server.tls_cluster) {
-                n->tls_port = atoi(port);
-            } else {
-                n->tcp_port = atoi(port);
-            }
-        } else if (!aux_tcp_port) {
-            n->tcp_port = atoi(port);
-        } else if (!aux_tls_port) {
-            n->tls_port = atoi(port);
-        }
-        /* In older versions of nodes.conf the "@busport" part is missing.
-         * In this case we set it to the default offset of 10000 from the
-         * base port. */
-        n->cport = busp ? atoi(busp) : (getNodeDefaultClientPort(n) + CLUSTER_PORT_INCR);
-
-        /* The plaintext port for client in a TLS cluster (n->pport) is not
-         * stored in nodes.conf. It is received later over the bus protocol. */
-
-        sdsfreesplitres(aux_argv, aux_argc);
-
-        /* Parse flags */
-        p = s = argv[2];
-        while (p) {
-            p = strchr(s, ',');
-            if (p) *p = '\0';
-            if (!strcasecmp(s, "myself")) {
-                serverAssert(server.cluster->myself == NULL);
-                myself = server.cluster->myself = n;
-                n->flags |= CLUSTER_NODE_MYSELF;
-            } else if (!strcasecmp(s, "master") || !strcasecmp(s, "primary")) {
-                n->flags |= CLUSTER_NODE_PRIMARY;
-            } else if (!strcasecmp(s, "slave") || !strcasecmp(s, "replica")) {
-                n->flags |= CLUSTER_NODE_REPLICA;
-            } else if (!strcasecmp(s, "fail?")) {
-                n->flags |= CLUSTER_NODE_PFAIL;
-            } else if (!strcasecmp(s, "fail")) {
-                n->flags |= CLUSTER_NODE_FAIL;
-                LEGACY_DATA(n)->fail_time = mstime();
-            } else if (!strcasecmp(s, "handshake")) {
-                n->flags |= CLUSTER_NODE_HANDSHAKE;
-            } else if (!strcasecmp(s, "noaddr")) {
-                n->flags |= (CLUSTER_NODE_NOADDR | CLUSTER_NODE_FAIL);
-                LEGACY_DATA(n)->fail_time = mstime();
-            } else if (!strcasecmp(s, "nofailover")) {
-                n->flags |= CLUSTER_NODE_NOFAILOVER;
-            } else if (!strcasecmp(s, "noflags")) {
-                /* nothing to do */
-            } else {
-                serverPanic("Unknown flag in %s cluster config file", SERVER_TITLE);
-            }
-            if (p) s = p + 1;
-        }
-
-        /* Get primary if any. Set the primary and populate primary's
-         * replica list. */
-        if (argv[3][0] != '-') {
-            if (verifyClusterNodeId(argv[3], sdslen(argv[3])) == C_ERR) {
-                sdsfreesplitres(argv, argc);
-                goto fmterr;
-            }
-            primary = clusterLookupNode(argv[3], sdslen(argv[3]));
-            if (!primary) {
-                primary = createClusterNode(argv[3], 0);
-                clusterAddNode(primary);
-            }
-            /* shard_id can be absent if we are loading a nodes.conf generated
-             * by an older version; we should follow the primary's
-             * shard_id in this case */
-            if (auxFieldHandlers[af_shard_id].isPresent(n) == 0) {
-                memcpy(n->shard_id, primary->shard_id, CLUSTER_NAMELEN);
-                clusterAddNodeToShard(primary->shard_id, n);
-            } else if (clusterGetNodesInMyShard(primary) != NULL &&
-                       memcmp(primary->shard_id, n->shard_id, CLUSTER_NAMELEN) != 0) {
-                /* If the primary has been added to a shard and this replica has
-                 * a different shard id stored in nodes.conf, update it to match
-                 * the primary instead of aborting the startup. */
-                serverLog(LL_NOTICE,
-                          "Node %.40s has a different shard id (%.40s) than its primary %.40s (%.40s). "
-                          "Updating replica's shard id to match primary's shard id.",
-                          n->name, n->shard_id, primary->name, primary->shard_id);
-                clusterRemoveNodeFromShard(n);
-                memcpy(n->shard_id, primary->shard_id, CLUSTER_NAMELEN);
-                clusterAddNodeToShard(primary->shard_id, n);
-            }
-            n->replicaof = primary;
-            clusterNodeAddReplica(primary, n);
-        } else if (auxFieldHandlers[af_shard_id].isPresent(n) == 0) {
-            /* n is a primary but it does not have a persisted shard_id.
-             * This happens if we are loading a nodes.conf generated by
-             * an older version of the server. We should manually update the
-             * shard membership in this case */
-            clusterAddNodeToShard(n->shard_id, n);
-        }
-
-        /* Set ping sent / pong received timestamps */
-        if (atoi(argv[4])) LEGACY_DATA(n)->ping_sent = mstime();
-        if (atoi(argv[5])) LEGACY_DATA(n)->pong_received = mstime();
-
-        /* Set configEpoch for this node.
-         * If the node is a replica, set its config epoch to 0.
-         * If it's a primary, load the config epoch from the configuration file. */
-        LEGACY_DATA(n)->configEpoch = (nodeIsReplica(n) && n->replicaof) ? 0 : strtoull(argv[6], NULL, 10);
-
-        /* Populate hash slots served by this instance. */
-        for (j = 8; j < argc; j++) {
-            int start, stop;
-
-            if (argv[j][0] == '[') {
-                /* Here we handle migrating / importing slots */
-                int slot;
-                char direction;
-                clusterNode *cn;
-
-                p = strchr(argv[j], '-');
-                serverAssert(p != NULL);
-                *p = '\0';
-                direction = p[1]; /* Either '>' or '<' */
-                slot = atoi(argv[j] + 1);
-                if (slot < 0 || slot >= CLUSTER_SLOTS) {
-                    sdsfreesplitres(argv, argc);
-                    goto fmterr;
-                }
-                p += 3;
-
-                char *pr = strchr(p, ']');
-                size_t node_len = pr - p;
-                if (pr == NULL || verifyClusterNodeId(p, node_len) == C_ERR) {
-                    sdsfreesplitres(argv, argc);
-                    goto fmterr;
-                }
-                cn = clusterLookupNode(p, CLUSTER_NAMELEN);
-                if (!cn) {
-                    cn = createClusterNode(p, 0);
-                    clusterAddNode(cn);
-                }
-                if (direction == '>') {
-                    setMigratingSlotDest(slot, cn);
-                } else {
-                    setImportingSlotSource(slot, cn);
-                }
-                continue;
-            } else if ((p = strchr(argv[j], '-')) != NULL) {
-                *p = '\0';
-                start = atoi(argv[j]);
-                stop = atoi(p + 1);
-            } else {
-                start = stop = atoi(argv[j]);
-            }
-            if (start < 0 || start >= CLUSTER_SLOTS || stop < 0 || stop >= CLUSTER_SLOTS) {
-                sdsfreesplitres(argv, argc);
-                goto fmterr;
-            }
-            while (start <= stop) clusterAddSlot(n, start++);
-        }
-
-        sdsfreesplitres(argv, argc);
-    }
-    /* Config sanity check */
-    if (server.cluster->myself == NULL) goto fmterr;
-
-    zfree(line);
-    fclose(fp);
-    serverAssert(tmp_cluster_nodes != NULL);
-    dictRelease(tmp_cluster_nodes);
-
-    serverLog(LL_NOTICE, "Node configuration loaded, I'm %.40s", myself->name);
-
-    /* Something that should never happen: currentEpoch smaller than
-     * the max epoch found in the nodes configuration. However we handle this
-     * as some form of protection against manual editing of critical files. */
-    if (clusterGetMaxEpoch() > LEGACY_STATE()->currentEpoch) {
-        LEGACY_STATE()->currentEpoch = clusterGetMaxEpoch();
-    }
-    return C_OK;
-
-fmterr:
-    serverPanic("Unrecoverable error: corrupted cluster config file \"%s\".", line);
-}
-
-/* Cluster node configuration is exactly the same as CLUSTER NODES output.
- *
- * This function writes the node config and returns C_OK, on error C_ERR
- * is returned.
- *
- * Note: we need to write the file in an atomic way from the point of view
- * of the POSIX filesystem semantics, so that if the server is stopped
- * or crashes during the write, we'll end with either the old file or the
- * new one. Since we have the full payload to write available we can use
- * a single write to write the whole file. If the pre-existing file was
- * bigger we pad our payload with newlines that are anyway ignored and truncate
- * the file afterward. */
-int clusterSaveConfig(int do_fsync) {
-    sds ci, tmpfilename;
-    size_t content_size, offset = 0;
-    ssize_t written_bytes;
-    int fd = -1;
-    int retval = C_ERR;
-    mstime_t latency;
-
-    LEGACY_STATE()->todo_before_sleep &= ~CLUSTER_TODO_SAVE_CONFIG;
-
-    /* Get the nodes description and concatenate our "vars" directive to
-     * save currentEpoch and lastVoteEpoch. */
-    ci = clusterGenNodesDescription(NULL, CLUSTER_NODE_HANDSHAKE, 0);
-    ci = sdscatfmt(ci, "vars currentEpoch %U lastVoteEpoch %U\n",
-                   (unsigned long long)LEGACY_STATE()->currentEpoch,
-                   (unsigned long long)LEGACY_STATE()->lastVoteEpoch);
-    content_size = sdslen(ci);
-
-    /* Create a temp file with the new content. */
-    tmpfilename = sdscatfmt(sdsempty(), "%s.tmp-%i-%I", server.cluster_configfile, (int)getpid(), mstime());
-    latencyStartMonitor(latency);
-    if ((fd = open(tmpfilename, O_WRONLY | O_CREAT, 0644)) == -1) {
-        serverLog(LL_WARNING, "Could not open temp cluster config file: %s", strerror(errno));
-        goto cleanup;
-    }
-    latencyEndMonitor(latency);
-    latencyAddSampleIfNeeded("cluster-config-open", latency);
-    latencyTraceIfNeeded(cluster, cluster_config_open, latency);
-    latencyStartMonitor(latency);
-    while (offset < content_size) {
-        written_bytes = write(fd, ci + offset, content_size - offset);
-        if (written_bytes <= 0) {
-            if (errno == EINTR) continue;
-            serverLog(LL_WARNING, "Failed after writing (%zd) bytes to tmp cluster config file: %s", offset,
-                      strerror(errno));
-            goto cleanup;
-        }
-        offset += written_bytes;
-    }
-    latencyEndMonitor(latency);
-    latencyAddSampleIfNeeded("cluster-config-write", latency);
-    latencyTraceIfNeeded(cluster, cluster_config_write, latency);
-    if (do_fsync) {
-        latencyStartMonitor(latency);
-        LEGACY_STATE()->todo_before_sleep &= ~CLUSTER_TODO_FSYNC_CONFIG;
-        if (valkey_fsync(fd) == -1) {
-            serverLog(LL_WARNING, "Could not sync tmp cluster config file: %s", strerror(errno));
-            goto cleanup;
-        }
-        latencyEndMonitor(latency);
-        latencyAddSampleIfNeeded("cluster-config-fsync", latency);
-        latencyTraceIfNeeded(cluster, cluster_config_fsync, latency);
-    }
-
-    latencyStartMonitor(latency);
-    if (rename(tmpfilename, server.cluster_configfile) == -1) {
-        serverLog(LL_WARNING, "Could not rename tmp cluster config file: %s", strerror(errno));
-        goto cleanup;
-    }
-    latencyEndMonitor(latency);
-    latencyAddSampleIfNeeded("cluster-config-rename", latency);
-    latencyTraceIfNeeded(cluster, cluster_config_rename, latency);
-    if (do_fsync) {
-        latencyStartMonitor(latency);
-        if (fsyncFileDir(server.cluster_configfile) == -1) {
-            serverLog(LL_WARNING, "Could not sync cluster config file dir: %s", strerror(errno));
-            goto cleanup;
-        }
-        latencyEndMonitor(latency);
-        latencyAddSampleIfNeeded("cluster-config-dir-fsync", latency);
-        latencyTraceIfNeeded(cluster, cluster_config_dir_fsync, latency);
-    }
-    retval = C_OK; /* If we reached this point, everything is fine. */
-
-cleanup:
-    if (fd != -1) {
-        latencyStartMonitor(latency);
-        close(fd);
-        latencyEndMonitor(latency);
-        latencyAddSampleIfNeeded("cluster-config-close", latency);
-        latencyTraceIfNeeded(cluster, cluster_config_close, latency);
-    }
-    if (retval == C_ERR) {
-        latencyStartMonitor(latency);
-        unlink(tmpfilename);
-        latencyEndMonitor(latency);
-        latencyAddSampleIfNeeded("cluster-config-unlink", latency);
-        latencyTraceIfNeeded(cluster, cluster_config_unlink, latency);
-    }
-    sdsfree(tmpfilename);
-    sdsfree(ci);
-    return retval;
-}
-
-/* Save the cluster configuration file. If the save fails, exit the process. */
-void clusterSaveConfigOrDie(int do_fsync) {
-    if (clusterSaveConfig(do_fsync) == C_ERR) {
-        serverLog(LL_WARNING, "Fatal: can't update cluster config file.");
-        exit(1);
-    }
-}
-
-/* Save the cluster configuration file. If the save fails, print the log. */
-#define CONFIG_SAVE_LOG_ERROR_RATE 30 /* Seconds between errors logging. */
-void clusterSaveConfigOrLog(int do_fsync) {
-    if (clusterSaveConfig(do_fsync) == C_ERR) {
-        static time_t last_save_error_log = 0;
-        /* Limit logging rate to 1 line per CONFIG_SAVE_LOG_ERROR_RATE seconds. */
-        if ((server.unixtime - last_save_error_log) > CONFIG_SAVE_LOG_ERROR_RATE) {
-            serverLog(LL_WARNING, "Cluster config updated even though writing "
-                                  "the cluster config file to disk failed.");
-            last_save_error_log = server.unixtime;
-        }
-    }
-}
-
-/* Lock the cluster config using flock(), and retain the file descriptor used to
- * acquire the lock so that the file will be locked as long as the process is up.
- *
- * This works because we always update nodes.conf with a new version
- * in-place, reopening the file, and writing to it in place (later adjusting
- * the length with ftruncate()).
- *
- * On success C_OK is returned, otherwise an error is logged and
- * the function returns C_ERR to signal a lock was not acquired. */
-int clusterLockConfig(char *filename) {
-/* flock() does not exist on Solaris
- * and a fcntl-based solution won't help, as we constantly re-open that file,
- * which will release _all_ locks anyway
- */
-#if !defined(__sun)
-    /* To lock it, we need to open the file in a way it is created if
-     * it does not exist, otherwise there is a race condition with other
-     * processes. */
-    int fd = open(filename, O_WRONLY | O_CREAT | O_CLOEXEC, 0644);
-    if (fd == -1) {
-        serverLog(LL_WARNING, "Can't open %s in order to acquire a lock: %s", filename, strerror(errno));
-        return C_ERR;
-    }
-
-    if (flock(fd, LOCK_EX | LOCK_NB) == -1) {
-        if (errno == EWOULDBLOCK) {
-            serverLog(LL_WARNING,
-                      "Sorry, the cluster configuration file %s is already used "
-                      "by a different Cluster node. Please make sure that "
-                      "different nodes use different cluster configuration "
-                      "files.",
-                      filename);
-        } else {
-            serverLog(LL_WARNING, "Impossible to lock %s: %s", filename, strerror(errno));
-        }
-        close(fd);
-        return C_ERR;
-    }
-    /* Lock acquired: leak the 'fd' by not closing it until shutdown time, so that
-     * we'll retain the lock to the file as long as the process exists.
-     *
-     * After fork, the child process will get the fd opened by the parent process,
-     * we need save `fd` to `cluster_config_file_lock_fd`, so that in serverFork(),
-     * it will be closed in the child process.
-     * If it is not closed, when the main process is killed -9, but the child process
-     * (valkey-aof-rewrite) is still alive, the fd(lock) will still be held by the
-     * child process, and the main process will fail to get lock, means fail to start. */
-    server.cluster_config_file_lock_fd = fd;
-#else
-    UNUSED(filename);
-#endif /* __sun */
-
-    return C_OK;
-}
-
 /* Some flags (currently just the NOFAILOVER flag) may need to be updated
  * in the "myself" node based on the current configuration of the node,
  * that may change at runtime via CONFIG SET. This function changes the
@@ -1401,6 +845,7 @@ static void clusterLegacyInit(void) {
         clusterAddNodeToShard(myself->shard_id, myself);
         saveconf = 1;
     }
+    myself = server.cluster->myself;
     if (saveconf) clusterSaveConfigOrDie(1);
 
     /* Port sanity check II
