+++ src/cluster_state.h
@@ -121,6 +121,8 @@ int clusterPrimariesHaveReplicas(void);
 int clusterNodeClearSlotBit(clusterNode *n, int slot);
 void clusterRemoveNodeFromShard(clusterNode *node);
 void clusterAddNodeToShard(const char *shard_id, clusterNode *node);
+void clusterAddNode(clusterNode *node);
+list *clusterGetNodesInMyShard(clusterNode *node);
 
 /* Node creation. */
 clusterNode *createClusterNode(char *nodename, int flags);
